requirements.txt
aiohttp

pandas
numpy
openpyxl

scikit-learn
xgboost
joblib
imbalanced-learn

shap
matplotlib
seaborn

scipy

google-generativeai
